Refactored AuthForm and fixed an issue whewre userObj didn't set to null when log out.
