+++ src/components/App.js
@@ -13,6 +13,8 @@ export default function App() {
           uid: user.uid,
           updateProfile: args => user.updateProfile(args),
         });
+      } else {
+        setUserObj(null);
       }
       setIsLoaded(true);
     });
